+++ demo/guihello/fontunits.h
@@ -1,0 +1,5 @@
+/* Font units */
+
+#include <gui/gui.hxx>
+
+Panel *font_units_demo(void);