+++ demo/guihello/guihello.c
@@ -6,6 +6,7 @@
 #include "sliders.h"
 #include "editor.h"
 #include "fontx.h"
+#include "fontunits.h"
 #include "form.h"
 #include "seltext.h"
 #include "popcom.h"
@@ -139,6 +140,9 @@ static void i_set_panel(App *app, const uint32_t index)
         panel = font_x_scale();
         break;
     case 30:
+        panel = font_units_demo();
+        break;
+    case 31:
         panel = reduce(app->window);
         break;
     }
@@ -200,6 +204,7 @@ static Panel *i_panel(App *app)
     listbox_add_elem(list, "Dynamic menus", NULL);
     listbox_add_elem(list, "IP Input", NULL);
     listbox_add_elem(list, "Font x-scale", NULL);
+    listbox_add_elem(list, "Font units", NULL);
     listbox_add_elem(list, "Reduce components", NULL);
     listbox_select(list, 0, TRUE);
     listbox_OnSelect(list, listener(app, i_OnSelect, App));