+++ src/osgui/osx/ostext.m
@@ -50,6 +50,7 @@
     char_t ffamily[64];
     real32_t fsize;
     uint32_t fstyle;
+    uint32_t funits;
     align_t palign;
     real32_t pspacing;
     real32_t pafter;
@@ -351,6 +352,7 @@ OSText *ostext_create(const uint32_t flags)
     view->ffamily[0] = '\0';
     view->fsize = REAL32_MAX;
     view->fstyle = UINT32_MAX;
+    view->funits = UINT32_MAX;
     view->palign = ENUM_MAX(align_t);
     view->pspacing = REAL32_MAX;
     view->pafter = REAL32_MAX;
@@ -535,7 +537,7 @@ void ostext_set_rtf(OSText *view, Stream *rtf_in)
 
 /*---------------------------------------------------------------------------*/
 
-static NSFont *i_font_create(const char_t *family, const real32_t size, const uint32_t style)
+static NSFont *i_font_create(const char_t *family, const real32_t size, const uint32_t style, const uint32_t units)
 {
     NSFont *nsfont = nil;
     cassert(size > 0.f);
@@ -552,6 +554,17 @@ static NSFont *i_font_create(const char_t *family, const real32_t size, const ui
 
     {
         Font *font = font_create(family, size, style);
+
+        if ((units & ekFCELL) == ekFCELL)
+        {
+            Font *nfont = NULL;
+            real32_t w, h;
+            font_extents(font, "REFTEXT", -1, &w, &h);
+            nfont = font_create(family, size * (size / h), style);
+            font_destroy(&font);
+            font = nfont;
+        }
+
         nsfont = cast(font_native(font), NSFont);
         font_destroy(&font);
     }
@@ -565,7 +578,7 @@ static void i_change_font(OSXTextView *lview)
 {
     NSFont *font = nil;
     cassert_no_null(lview);
-    font = i_font_create(lview->ffamily, lview->fsize, lview->fstyle);
+    font = i_font_create(lview->ffamily, lview->fsize, lview->fstyle, lview->funits);
     if (font != nil)
     {
         NSNumber *under = (lview->fstyle & ekFUNDERLINE) ? [NSNumber numberWithInt:NSUnderlineStyleSingle] : [NSNumber numberWithInt:NSUnderlineStyleNone];
@@ -673,6 +686,11 @@ void ostext_property(OSText *view, const gui_text_t param, const void *value)
         break;
 
     case ekGUI_TEXT_UNITS:
+        if (lview->funits != *cast(value, uint32_t))
+        {
+            lview->funits = *cast(value, uint32_t);
+            i_change_font(lview);
+        }
         break;
 
     case ekGUI_TEXT_SIZE: