+++ src/sewer/bmath.hpp
@@ -11,6 +11,9 @@
 
 /* Math funcions */
 
+#ifndef __BMATH_HPP__
+#define __BMATH_HPP__
+
 #include "sewer.hxx"
 
 template < typename real >
@@ -74,3 +77,5 @@ struct BMath
     _sewer_api static const real kRAD2DEG;
     _sewer_api static const real kINFINITY;
 };
+
+#endif