+++ Changelog.md
@@ -9,7 +9,7 @@
 - `button_get_text()`. [Commit](https://github.com/frang75/nappgui_src/commit/c57f87ff5916ca624bf5bd93484b68f693c7eb25).
 - `button_get_image()`. [Commit](https://github.com/frang75/nappgui_src/commit/7d158af6719910c4c89425830ac55bc6d65d1e17).
 - `button_get_image_alt()`. [Commit](https://github.com/frang75/nappgui_src/commit/7d158af6719910c4c89425830ac55bc6d65d1e17).
-- New functions in `PopUp` and `Combo`. [Doc](https://nappgui.com/en/gui/combo.html). [Commit]().
+- New functions in `PopUp` and `Combo`. [Doc](https://nappgui.com/en/gui/combo.html). [Commit](https://github.com/frang75/nappgui_src/commit/850954a50d2dfde51cd846dbb306cd9c6345c2dc).
     - `popup_ins_elem()`.
     - `popup_del_elem()`.
     - `combo_OnFocus()`.
@@ -26,12 +26,15 @@
     - `combo_list_height()`.
     - `combo_selected()`.
     - `combo_get_selected()`.
+- `font_units()`. [Commit]().
+- Support for `ekFCELL` in `textview_units()`. [Commit]().
 
 ### Fixed
 
 - Issue in `label_align()`. [Issue](https://github.com/frang75/nappgui_src/issues/225). [Commit](https://github.com/frang75/nappgui_src/commit/0afe4f04611fc14ce8274bcb2f200660f3745f0e).
 - Issue in `Edit` single-line scroll in macOS. [Issue](https://github.com/frang75/nappgui_src/issues/230). [Commit](https://github.com/frang75/nappgui_src/commit/9f54924b9b0cc1f405a13d47a6b305a8f69f5fc0).
-- Issues in `Combo`. [Issue](https://github.com/frang75/nappgui_src/issues/86). [Issue](https://github.com/frang75/nappgui_src/issues/81). [Issue](https://github.com/frang75/nappgui_src/issues/68). [Commit]().
+- Issues in `Combo`. [Issue](https://github.com/frang75/nappgui_src/issues/86). [Issue](https://github.com/frang75/nappgui_src/issues/81). [Issue](https://github.com/frang75/nappgui_src/issues/68). [Commit](https://github.com/frang75/nappgui_src/commit/850954a50d2dfde51cd846dbb306cd9c6345c2dc).
+- Added include guard to `bmath.hpp`. [Issue](https://github.com/frang75/nappgui_src/issues/235). [Commit]().
 
 ### Changed